refactoring of code for copying JMS properties, unit test case and log4j - testing in progress
