+++ src/test/resources/log4j.properties
@@ -14,9 +14,7 @@
 #  limitations under the License.
 #
 log4j.rootLogger=DEBUG, stdout
-
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
 log4j.appender.stdout.layout.ConversionPattern=[%d] %p %m (%c:%L)%n
-
 log4j.logger.org.apache.kafka=ERROR
